requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
psycopg[binary]
Werkzeug==2.2.2
python-dotenv
